move selection modification into new usecase
