--- src/main/groovy/life/qbic/portal/sampletracking/app/inspectsample/RetrieveLocation.groovy
@@ -1,13 +1,0 @@
-package life.qbic.portal.sampletracking.app.inspectsample
-
-import life.qbic.datamodel.samples.Location
-
-
-class RetrieveLocation {
-
-    List<Location> availableLocations(){}
-
-    Location currentLocationForSample(String sampleId) {
-
-    }
-}
