--- src/main/groovy/life/qbic/portal/sampletracking/app/inspectsample/ValidateQBiCId.groovy
@@ -1,8 +1,0 @@
-package life.qbic.portal.sampletracking.app.inspectsample
-
-class ValidateQBiCId {
-
-    boolean isValidId(String id) {
-
-    }
-}
