+++ src/main/groovy/life/qbic/portal/sampletracking/app/samples/selection/ModifySampleSelection.groovy
@@ -1,0 +1,4 @@
+package life.qbic.portal.sampletracking.app.samples.selection
+
+class ModifySampleSelection {
+}
