+++ src/main/groovy/life/qbic/portal/sampletracking/app/samples/selection/ModifySampleSelectionOutput.groovy
@@ -1,9 +1,9 @@
-package life.qbic.portal.sampletracking.ui
+package life.qbic.portal.sampletracking.app.samples.selection
 
 import life.qbic.datamodel.samples.Sample
 
-interface ListViewModel {
+interface ModifySampleSelectionOutput {
     void addSamples(Sample... samples)
     void removeSamples(Sample... samples)
-    void clear()
+    void clearSamples()
 }
