+++ src/main/resources/life/qbic/portal/sampletracking/AppWidgetSet.gwt.xml
@@ -12,5 +12,5 @@
 
     
 
-    <inherits name="life.qbic.portal.portlet.AppWidgetSet" />
+    
 </module>
